requirements.txt
requests
python-dotenv
pydantic
pandas
boto3